requirements.txt
anthropic==0.25.9
huggingface-hub==0.23.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1687808433287/work
openai==0.28.0
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1688740593370/work
python-dotenv==1.0.1
scikit-learn==1.3.2
scipy==1.10.1